Refactor AppRoot to lazy module with v8_middleware and appRootMiddleware
- Extract RouteErrorBoundary to its own module for reuse
- Convert AppRoot from a static element to a lazy-loaded route module
  (exports Component, loader, ErrorBoundary instead of default export)
- Enable v8_middleware in createBrowserRouter; add appRootDataContext to
  the per-navigation router context so middleware can share data with loaders
- Add appRootMiddleware on the AppRoot route: runs the AppRootQuery once
  per navigation and stores the result in appRootDataContext, so the query
  is never duplicated across middleware and loaders
- AppRoot loader now reads from appRootDataContext rather than issuing its
  own query; retains the profile-setup mutation and redirect logic
- Replace EditEventGuard (render-time side-effect guard) with
  editEventMiddleware: reads siteMode from appRootDataContext and redirects
  single-event sites to /admin_events before any rendering occurs; restore
  LoginRequiredRouteGuard as the route element
- Fix EventPageGuard: move navigate() call into useEffect to avoid
  calling a side effect during render
